Created a new function for scanning in arbitrary directions
This new function is to be used in the scan_winner function
and its purpose is to replace the four other functions that
are currently being used to scan in all of the different
directions.

I have applied replaced the code that scans columns succesfully.
Now I need to do the same for the rest of the functions.